Simplify timeline architecture: remove dead code, consolidate hooks and utils
- Delete 7 unused files: useTimelineChanges, useCategoryHeights,
  TimelineLabels, TimelineCategoryEvents, TimelineContainer (x2),
  timelineRange
- Create useTimelineState hook composing useEvents, useTimelineTitle,
  useCategories, and useTimelineScale into a single entry point
- Merge timelineRange.ts into dateUtils.ts, remove unused exports
  from timelineUtils.ts
- Deduplicate: shared debounce in useTimelineScroll, shared
  formatDateForCSV in excelExport, extract eventContent JSX in
  TimelineEvent
- Inline TimelineContainer wrapper directly into App.tsx
- Update TIMELINE_ARCHITECTURE.md to reflect changes
